hasClassUsage method simplified
