+++ src/main/kotlin/mappers/ASTUtil.kt
@@ -1,9 +1,14 @@
 package mappers
 
+import SourceClassInfo
+import com.github.javaparser.ast.CompilationUnit
+import com.github.javaparser.ast.Node
 import com.github.javaparser.ast.NodeList
 import com.github.javaparser.ast.expr.*
 import org.jetbrains.kotlin.com.intellij.psi.impl.source.tree.LeafPsiElement
+import org.jetbrains.kotlin.name.parentOrNull
 import org.jetbrains.kotlin.psi.KtAnnotationEntry
+import org.jetbrains.kotlin.psi.KtFile
 import org.jetbrains.kotlin.psi.ValueArgument
 import org.jetbrains.kotlin.psi.psiUtil.findDescendantOfType
 
@@ -46,3 +51,51 @@ private fun KtAnnotationEntry.getValue(): String? =
     this.valueArguments.find {
         it.getArgumentName().let { name -> name == null || name.asName.asString() == "value" }
     } ?.toUnescapedString()
+
+fun Node.hasClassUsage(sourceClass: SourceClassInfo, requirePackage: Boolean): Boolean {
+    return when (this) {
+        is ObjectCreationExpr -> {
+            this.type.nameAsString == sourceClass.name &&
+                    (!requirePackage || sourceClass.pkg == this.type.scope.orElse(null)?.nameAsString)
+        }
+
+        is MethodCallExpr -> {
+            val fieldAccessExpr = this.scope.orElse(null) as? FieldAccessExpr
+            fieldAccessExpr?.let {
+                it.nameAsString == sourceClass.name &&
+                        (!requirePackage || sourceClass.pkg == it.scope?.toString())
+            } ?: false
+        }
+
+        else -> false
+    }
+}
+
+fun CompilationUnit.hasImport(fqcn: String, packageName: String): Boolean = this.imports.any { import ->
+    if (import.isStatic) {
+        if (import.isAsterisk) {
+            // import static org.test.SourceClass.*
+            fqcn == import.name.asString()
+        } else {
+            // import static org.test.SourceClass.doSmth
+            fqcn == (import.name.qualifier.orElse(null)?.asString() ?: "")
+        }
+    } else if (!import.isAsterisk) {
+        // import org.test.SourceClass
+        fqcn == import.name.asString()
+    } else {
+        // import org.test.*
+        packageName == (import.name.qualifier.orElse(null)?.asString() ?: "")
+    }
+}
+fun KtFile.hasImport(fqcn: String, packageName: String): Boolean = this.importDirectives.any {
+    val importedFqName = it.importedFqName?.asString()
+    if (it.isAllUnder) {
+        // import org.test.SourceClass.* or import org.test.*
+        fqcn == importedFqName || packageName == importedFqName
+    } else { // import org.test.SourceClass or import org.test.SourceClass.doSmth
+        fqcn == importedFqName || fqcn == it.importedFqName?.parentOrNull()?.asString()
+    }
+}
+
+
