--- src/main/kotlin/mappers/ClassUsageResolver.kt
@@ -1,7 +1,0 @@
-package mappers
-
-import SourceClassInfo
-
-interface ClassUsageResolver<File, Cls> {
-    fun isSourceClassUsed(testFile: File, testClass: Cls, sourceClass: SourceClassInfo): Boolean
-}
